Changing sensor advertisement from direct to fanout exchange
This commit changes the way sensor gets advertised to consumers.
No instead of sending advertisement message to direct exchange,
which is also default exchange, sensor is sending message to
fanout exchange. Fanout exchange doesn't need queue name to
determin where the message goes and it will send the message to
every copy of the queue bound to exchange.
It's up to the consumer to create message queue.